🎨 component refacturbation
