+++ src/components/ColouredContainer.js
@@ -1,7 +1,23 @@
 import React, { useState, useRef, useEffect } from "react";
-import { Container } from "./styles";
+import styled from "styled-components";
+import { rgba } from "polished";
 
-export const ColouredContainer = ({ children, ...props }) => {
+const Container = styled.div.attrs(
+  ({ clientX, clientY, clientWidth, clientHeight, style }) => ({
+    style: {
+      backgroundColor: rgba(
+        Math.round((clientX / clientHeight) * 255),
+        255 - Math.round((clientY / clientWidth) * 255),
+        255 - Math.round((clientX / clientHeight) * 255),
+        0.75
+      )
+    }
+  })
+)`
+  transition: background-color 0.2s linear;
+`;
+
+const RawColouredContainer = ({ children, ...props }) => {
   const myRef = useRef();
   const [clientX, setClientX] = useState(0);
   const [clientY, setClientY] = useState(0);
@@ -44,3 +60,5 @@ export const ColouredContainer = ({ children, ...props }) => {
     </Container>
   );
 };
+
+export const ColouredContainer = styled(RawColouredContainer)``;
